librarymanagement/src/test/java/com/mo/librarymanagement/service/MyUserDetailsServiceTest.java
package com.mo.librarymanagement.service;

import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.times;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.security.crypto.password.PasswordEncoder;

import com.mo.librarymanagement.entity.User;
import com.mo.librarymanagement.repository.UserRepository;

class MyUserDetailsServiceTest {

    @Mock
    private UserRepository userRepository;

    @Mock
    private PasswordEncoder passwordEncoder;

    @InjectMocks
    private MyUserDetailsService myUserDetailsService;

    @BeforeEach
    public void setUp() {
        MockitoAnnotations.openMocks(this);
    }

    @Test
    public void testLoadUserByUsername_UserFound() {
        User mockUser = new User();
        mockUser.setUsername("user");
        mockUser.setPassword("password");
        when(userRepository.findByUsername("user")).thenReturn(mockUser);

        org.springframework.security.core.userdetails.UserDetails userDetails = 
            myUserDetailsService.loadUserByUsername("user");

        assertNotNull(userDetails);
        assertEquals("user", userDetails.getUsername());
        assertEquals("password", userDetails.getPassword());
    }

    @Test
    public void testLoadUserByUsername_UserNotFound() {
        when(userRepository.findByUsername("unknownUser")).thenReturn(null);

        assertThrows(UsernameNotFoundException.class, () -> {
            myUserDetailsService.loadUserByUsername("unknownUser");
        });
    }

    @Test
    public void testSaveUser_Success() {
        User user = new User();
        user.setUsername("user");
        user.setPassword("encodedPassword");

        when(passwordEncoder.encode("password")).thenReturn("encodedPassword");
        when(userRepository.save(any(User.class))).thenReturn(user);

        myUserDetailsService.saveUser("user", "password");

        verify(userRepository, times(1)).save(any(User.class));
    }
}